Tags support -------------
Hobix now support tags. Each entry may contain a field, named 'tags',
which contain any relevant tag for the current entry.
By default, a set of tags is infered by using the entry base path.
For example if the entry path is '/foo/bar/Baz.yaml', then the default
tags will be [ foo, bar ].
This behavior can be avoided, by adding the line

   Hobix::Entry.no_implicit_tags

to one of the included plugins.

If a tag should be set, by default, for all entries, one can add :

  class Hobix::Entry
    def force_keywords; [ 'my_default_tag' ] end
  end

to one of the included plugins.

If the 'hobix/out/tags' plugin is required, and a file, whose 1st stem
is tags (for example 'tags.html.quick') is found in the skel
directory, then a /tags directory will be created when the blog will
be generated, and for each known tags, a directory will be created,
with an index.html that old all of the relevant anetries for this tag.


Files changed:
--------------

commandline.rb : to allow the display of the default (infered) tags
                 field
entry.rb       : the 'tags' support by itself
rss.rb/atom.rb : use tags to feed 'subject'
filesys.rb     : added tags field, and set the entry id if it's
                 reloaded
tags.rb        : plugin added